fix(sharing): enforce access controls end to end
Harden API, collaboration, and web permission boundaries for shared content. Validate public links, imports, positions, persistence writers, destructive actions, and test setup while preserving derived metadata and observable error handling.

+++ packages/api/src/routes/export.ts
@@ -27,7 +27,7 @@ exportRoute.get('/export', async (c) => {
       from pages
       where is_deleted = false
         and id in (select page_id from get_accessible_page_ids($1))
-      order by parent_id nulls first, position asc
+      order by parent_id nulls first, position::numeric asc
     `,
     [user.id],
   );
